requirements.txt
pyusb==1.0.0b1
socketIO-client==0.4